refactor: 도메인 모델 패키지 경로 변경 및 MainActivity 삭제
