+++ app/src/main/java/woowacourse/shopping/data/repository/product/RemoteProductRepository.kt
@@ -2,8 +2,8 @@ package woowacourse.shopping.data.repository.product
 
 import woowacourse.shopping.data.datasource.product.ProductRemoteDataSource
 import woowacourse.shopping.data.mapper.toDomain
-import woowacourse.shopping.domain.product.Product
-import woowacourse.shopping.domain.product.Products
+import woowacourse.shopping.domain.model.product.Product
+import woowacourse.shopping.domain.model.product.Products
 import woowacourse.shopping.domain.repository.ProductRepository
 
 class RemoteProductRepository(
