+++ app/src/main/java/woowacourse/shopping/domain/model/cart/Cart.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain.cart
+package woowacourse.shopping.domain.model.cart
 
 data class Cart(
     val cartItems: CartItems = CartItems(),
