+++ app/src/main/java/woowacourse/shopping/domain/model/cart/Quantity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain.cart
+package woowacourse.shopping.domain.model.cart
 
 @JvmInline
 value class Quantity(
