+++ app/src/main/java/woowacourse/shopping/domain/model/product/Price.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain.product
+package woowacourse.shopping.domain.model.product
 
 @JvmInline
 value class Price(
