+++ app/src/main/java/woowacourse/shopping/domain/model/product/Products.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain.product
+package woowacourse.shopping.domain.model.product
 
 class Products(
     val items: List<Product>,
