+++ app/src/main/java/woowacourse/shopping/ui/cart/CartItemUiModel.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.ui.cart
 
-import woowacourse.shopping.domain.product.Product
+import woowacourse.shopping.domain.model.product.Product
 
 data class CartItemUiModel(
     val id: Int,
