+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -12,8 +12,8 @@ import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import woowacourse.shopping.domain.cart.CartItems
-import woowacourse.shopping.domain.product.Product
+import woowacourse.shopping.domain.model.cart.CartItems
+import woowacourse.shopping.domain.model.product.Product
 import woowacourse.shopping.domain.repository.CartRepository
 import woowacourse.shopping.domain.repository.ProductRepository
 import woowacourse.shopping.domain.repository.RecentProductRepository
