+++ app/src/main/java/woowacourse/shopping/ui/productDetail/ProductDetailViewModel.kt
@@ -9,8 +9,8 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import woowacourse.shopping.domain.cart.Quantity
-import woowacourse.shopping.domain.product.Product
+import woowacourse.shopping.domain.model.cart.Quantity
+import woowacourse.shopping.domain.model.product.Product
 import woowacourse.shopping.domain.repository.CartRepository
 import woowacourse.shopping.domain.repository.ProductRepository
 import woowacourse.shopping.domain.repository.RecentProductRepository
