+++ app/src/test/java/woowacourse/shopping/product/PriceTest.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.product
 
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Test
-import woowacourse.shopping.domain.product.Price
+import woowacourse.shopping.domain.model.product.Price
 
 class PriceTest {
     @Test
